requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.3
gunicorn==22.0.0
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
RPi.GPIO==0.7.1
werkzeug==3.0.2
zipp==3.18.1
